08 Jul 23

+++ src/com/begin/examples/Exception1.java
@@ -4,7 +4,7 @@ public class Exception1 {
     public static void main(String[] args) {
         System.out.println(divide(5,0));
     }
-    public static double divide(int a, int b) {
+    public static int divide(int a, int b) {
         if (b == 0) {
             throw new IllegalArgumentException("Cannot divide by Zero");
         }
